DatabaseCreation/Headline_and_contents/requirement.txt
tweepy
alembic
SQLAlchemy==1.3.16
sqlalchemy-utils==0.36.8
psycopg2
PyMySQL==0.9.2
mysql.connector